Refactor for the account filter
