+++ background/redux-slices/utils/account-filter-utils.ts
@@ -1,3 +1,10 @@
+export type FilterAccount = {
+  id: string
+  isEnabled: boolean
+  name: string
+  thumbnailURL?: string
+}
+
 export type AccountData = {
   address: string
   name: string
@@ -5,9 +12,9 @@ export type AccountData = {
 }
 
 export const getAdditionalDataForFilter = (
-  address: string,
+  id: string,
   accounts: AccountData[]
 ): { name?: string; thumbnailURL?: string } => {
-  const result = accounts.find((account) => account.address === address)
+  const result = accounts.find((account) => account.address === id)
   return result ? { name: result.name, thumbnailURL: result.avatarURL } : {}
 }
